fix create challenge with bad root folder

+++ app/challenges/challenges_helper.py
@@ -1,7 +1,7 @@
 import zipfile
 import os
 import json
-from app.global_helper import check_challenge_in_store, check_zip_structure
+from global_helper import check_challenge_in_store, check_zip_structure
 from fastapi import HTTPException
 from database_sqlite.models import Challenge
 from shutil import rmtree
@@ -9,6 +9,7 @@ from sqlalchemy.ext.asyncio import (
     async_sessionmaker,
     AsyncSession,
 )
+from os.path import exists
 
 f = open('configure.json')
 data = json.load(f)
@@ -49,17 +50,20 @@ async def extract_challenge(async_session: async_sessionmaker[AsyncSession], cha
     os.remove(temp_zip_path)
 
     if folder_name_error:
-        rmtree(f"{challenges_dir}/{challenge_name}")
+        if exists(f"{challenges_dir}/{challenge_name}"):
+            rmtree(f"{challenges_dir}/{challenge_name}")
         delete_challenge_by_title(async_session, challenge_title)
         raise HTTPException(status_code=422, detail=f'Invalid challenge folder name "{challenge_name}" - is not equal to challenge title "{challenge_title}"')
 
     if challenge_already_exist_error:
-        rmtree(f"{challenges_dir}/{challenge_name}")
+        if exists(f"{challenges_dir}/{challenge_name}"):
+            rmtree(f"{challenges_dir}/{challenge_name}")
         delete_challenge_by_title(async_session, challenge_title)
         raise HTTPException(status_code=422, detail=f'Challenge "{challenge_name}" already exist in store!')
 
     if structure_error:
-        rmtree(f"{challenges_dir}/{challenge_name}")
+        if exists(f"{challenges_dir}/{challenge_name}"):
+            rmtree(f"{challenges_dir}/{challenge_name}")
         delete_challenge_by_title(async_session, challenge_title)
         raise HTTPException(status_code=422, detail=f'Bad challenge structure! Challenge required files: {str(required_challenge_files)}')
